1. add | as a symbol to text special cases; 2. patch fix a few rules for preprocessing; 3. add rules to handle TNM; 4. fix rules in BIO generation to make code more robust
