+++ text_process/text_special_cases.py
@@ -1,5 +1,5 @@
 SYMBOLS = {',', '?', '!', ':', '\'', '"', '(', ')', ';', '@', '^', '^', '&', '&', '$', '$', '£',
-           '[', ']', '{', '}', '<', '>', '+', '-', "*", "#", "%", "=", "~", '/', "_"}
+           '[', ']', '{', '}', '<', '>', '+', '-', "*", "#", "%", "=", "~", '/', "_", "|"}
 
 BREAK_SYMBOLS = {
     ',', '?', '!', ':', ')', ';', ']', '}', '>', '[', '{', '(', '<',
